+++ libraries/attribute_tools.rb
@@ -18,18 +18,21 @@ class RealityForge #nodoc
       # Deep merge value into node using output path. If the value provided is nil then perform operation.
       #
       # = Parameters
-      # * +node+:: The node into which the results will be deep merged.
+      # * +root_element+:: The node or mash into which the results will be deep merged.
       # * +output_path+:: The path on the node on which to deep merge the results.
       # * +value+:: The value to merge.
-      def deep_merge(node, output_path, value)
+      def deep_merge(root_element, output_path, value)
         if value
-          existing = output_path.split('.').inject(node.override) { |element, key| element[key] }
+          existing =
+            output_path.nil? ?
+              root_element :
+              output_path.split('.').inject(root_element.respond_to?(:override) ? root_element.override : root_element) { |element, k| element.nil? ? nil : element[k] }
           if existing
-            results = ::Chef::Mixin::DeepMerge.deep_merge(value, existing.to_hash).to_hash
+            results = ::Chef::Mixin::DeepMerge.deep_merge(value, existing).to_hash
           else
             results = value.dup
           end
-          set_attribute_on_node(node, output_path, results)
+          set_attribute(root_element, output_path, results)
         end
       end
 
@@ -42,7 +45,7 @@ class RealityForge #nodoc
       # * +prefix+:: The prefix already traversed to get to root.
       def ensure_attribute(root_element, key, type = nil, prefix = nil)
         key_parts = key.split('.')
-        output_entry = key_parts[0...-1].inject(root_element) { |element, k| element.nil? ? nil : element[k] }
+        output_entry = key_parts[0...-1].inject(root_element.to_hash) { |element, k| element.nil? ? nil : element[k] }
         value = output_entry ? output_entry[key_parts.last] : nil
         label = prefix ? "#{prefix}.#{key}" : key
         raise "Attribute '#{label}' is missing" unless value